Minor cast adjustment which shouldn't really have any impact
DCO-1.1-Signed-off-by: Ellie <kittens@wobble.ninja>

+++ horse64/widechar.c
@@ -317,6 +317,6 @@ int utf32_to_utf8(
         totallen += inneroutlen;
         i++;
     }
-    if (out_len) *out_len = totallen;
+    if (out_len) *out_len = (int64_t)totallen;
     return 1;
 }
